@@ -1,20 +1,24 @@
 package ar.edu.unlam.tallerweb1.delivery;
 
-import ar.edu.unlam.tallerweb1.annotations.RequireAuth;
 import ar.edu.unlam.tallerweb1.domain.auth.IServicioAuth;
-import ar.edu.unlam.tallerweb1.domain.auth.ServicioAuth;
 import ar.edu.unlam.tallerweb1.domain.publicaciones.IServicioPublicacion;
+import ar.edu.unlam.tallerweb1.model.Imagen;
 import ar.edu.unlam.tallerweb1.model.Publicacion;
+import ar.edu.unlam.tallerweb1.model.Publicacion_favorito;
 import ar.edu.unlam.tallerweb1.model.Usuario;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.persistence.PersistenceException;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.util.List;
+import java.util.Optional;
 
 @Controller
 @RequestMapping("/home")
@@ -24,10 +28,24 @@ public class ControladorHome {
     private IServicioPublicacion servicioPublicacion;
     private IServicioAuth servicioAuth;
 
+
     @Autowired
-    public void ControladorHome(IServicioPublicacion servicioPublicacion, IServicioAuth servicioAuth) {
+    public void ControladorHome(IServicioPublicacion servicioPublicacion, IServicioAuth servicioAuth){
+
         this.servicioPublicacion = servicioPublicacion;
+
         this.servicioAuth = servicioAuth;
+
+    }
+
+    private ModelMap iniciarModel(){
+        this.userAuth = this.servicioAuth.getUsuarioAutenticado();
+
+        ModelMap m = new ModelMap();
+
+        m.put("usuario", this.userAuth);
+
+        return m;
     }
 
     @RequestMapping(path = "/", method = RequestMethod.GET)
@@ -35,19 +53,98 @@ public class ControladorHome {
         return new ModelAndView("redirect:/home/feed");
     }
 
-    @RequireAuth //! leeme e.e
-    @RequestMapping(path = "/feed", method = RequestMethod.GET)
-    public ModelAndView home(HttpSession session) {
+    @RequestMapping(path = "/feed",method = RequestMethod.GET)
+    public ModelAndView feed(@RequestParam(required = false) String response ) {
 
-        Usuario userAuth = servicioAuth.getUsuarioAutenticado(); //leemeeee
+        ModelMap model = iniciarModel();
 
-        ModelMap model = new ModelMap();
-        //solicitar publicaciones
-        List<Publicacion> publicaciones = servicioPublicacion.listarPublicaciones();
+        if(userAuth != null){
 
-        model.addAttribute("usuario", userAuth);
-        model.addAttribute("publicaciones", publicaciones);
-        return new ModelAndView("index-feed", model);
+            //solicitar publicaciones
+            List<Publicacion> publicaciones = servicioPublicacion.listarPublicacionesDisponibles();
+
+            model.put("target","feed");
+            model.put("publicaciones", publicaciones);
+            model.put("response_f", response);
+
+            return new ModelAndView("index-feed",model);
+        }
+
+        return new ModelAndView("redirect:/login");
 
     }
+
+    @RequestMapping(path = "/favoritos",method = RequestMethod.GET)
+    public ModelAndView favoritos(@RequestParam(required = false)String response) {
+
+        ModelMap model = iniciarModel();
+
+        if(userAuth != null){
+
+            //solicitar publicaciones
+            List<Publicacion_favorito> favoritos = servicioPublicacion.listarFavoritosDeUsuario(userAuth.getId());
+
+            model.put("target","favoritos");
+            model.put("publicaciones", favoritos);
+            model.put("response_f", response);
+            return new ModelAndView("index-favorites",model);
+        }
+
+        return new ModelAndView("redirect:/login");
+
+    }
+
+    @RequestMapping(path = "/mispublicaciones",method = RequestMethod.GET)
+    public ModelAndView misPublicaciones(@RequestParam(required = false) String pid, HttpSession session) {
+
+        ModelMap model = iniciarModel();
+
+        if(userAuth != null){
+
+            //solicitar publicaciones
+            List<Publicacion> publicaciones = servicioPublicacion.listarPublicacionesPorUsuarioId(userAuth.getId());
+
+            model.put("loader", pid);
+            model.put("target","mispublicaciones");
+            model.put("publicaciones", publicaciones);
+
+            return new ModelAndView("index-misposts",model);
+        }
+
+        return new ModelAndView("redirect:/login");
+
+    }
+
+    @RequestMapping(path = "/favoritos/agregar")
+    public ModelAndView agregarFavorito(@RequestParam Long pid) {
+
+        String response;
+
+        try{
+
+            this.servicioPublicacion.agregarFavorito(pid, this.userAuth);
+            response = "success";
+
+        }catch (PersistenceException err){
+            System.out.println(err.getMessage());
+            response = "exist";
+
+        }
+
+        return this.feed(response);
+
+    }
+
+    @RequestMapping(path = "/favoritos/eliminar")
+    public ModelAndView eliminarFavorito(@RequestParam Long pid) {
+
+        this.servicioPublicacion.eliminarFavorito(pid, this.userAuth);
+
+        String response = "removed";
+
+        return this.favoritos(response);
+
+    }
+
+
 }
