@@ -1,7 +1,10 @@
 package ar.edu.unlam.tallerweb1.delivery;
 
+import ar.edu.unlam.tallerweb1.domain.Mensajes.IServicioMensajes;
+import ar.edu.unlam.tallerweb1.domain.auth.IServicioAuth;
+import ar.edu.unlam.tallerweb1.domain.mascota.IServicioMascota;
 import ar.edu.unlam.tallerweb1.domain.publicaciones.IServicioPublicacion;
-import ar.edu.unlam.tallerweb1.model.Publicacion;
+import ar.edu.unlam.tallerweb1.model.Mascota;
 import ar.edu.unlam.tallerweb1.model.Usuario;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -9,76 +12,88 @@ import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
-import javax.servlet.http.HttpSession;
+
+import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
+
 
 @Controller
 @RequestMapping("/publicacion")
 public class ControladorPublicacion {
 
-    private IServicioPublicacion servicioPublicacion;
+    private final IServicioAuth servicioAuth;
+    private final IServicioPublicacion servicioPublicacion;
+    private final IServicioMascota servicioMascota;
+    private final IServicioMensajes servicioMesnaje;
+
+    private Usuario userAuth;
 
     @Autowired
-    public ControladorPublicacion(IServicioPublicacion servicioPublicacion) {
+    public ControladorPublicacion(IServicioPublicacion servicioPublicacion, IServicioMascota servicioMascota, IServicioMensajes servicioMensaje, IServicioAuth servicioAuth) {
         this.servicioPublicacion = servicioPublicacion;
+        this.servicioAuth = servicioAuth;
+        this.servicioMascota = servicioMascota;
+        this.servicioMesnaje = servicioMensaje;
     }
 
-    @RequestMapping(path = "/crear-publicacion", method = RequestMethod.POST)
-    public ModelAndView guardarPublicacion(@ModelAttribute("publicacion") DatosPublicacion datosPublicacion, HttpSession session) {
-        ModelMap model = new ModelMap();
-
-        Usuario usuario = (Usuario) session.getAttribute("usuarioAutenticado");
-
-        try{
-            Publicacion publicacion = new Publicacion();
-
-            publicacion.setTitulo(datosPublicacion.getTitulo());
-            publicacion.setCuerpo(datosPublicacion.getCuerpo());
-           // publicacion.setAutorId(usuario.getId());
-            publicacion.setMascotaId(datosPublicacion.getMascotaId());
-            publicacion.setFechaCreacion(datosPublicacion.getFechaCreacion());
-
-            servicioPublicacion.guardarPublicacion(publicacion);
-        } catch (Exception e) {
-            model.put("error", "No se pudo guardar la publicacion");
-        }
-
-        model.put("publicacion", datosPublicacion);
-        return new ModelAndView("index-feed", model);
+    private ModelMap iniciarModel() {
+        this.userAuth = this.servicioAuth.getUsuarioAutenticado();
+        ModelMap m = new ModelMap();
+        m.put("usuario", this.userAuth);
+        return m;
     }
 
-    @RequestMapping(path = "/publicaciones-feed", method = RequestMethod.GET)
-    public ModelAndView verPublicacionesFeed() {
-        ModelMap model = new ModelMap();
-
-        List<Publicacion> publicaciones = servicioPublicacion.listarPublicaciones();
-
-        model.put("publicaciones", publicaciones);
-
-        return new ModelAndView("index-feed", model);
-    }
-
-    @RequestMapping(path = "/mis-publicaciones", method = RequestMethod.GET)
-    public ModelAndView verPublicacionesPorUsuarioId(@ModelAttribute("datosPublicacion") DatosPublicacion datosPublicacion) {
-        ModelMap model = new ModelMap();
-        List<Publicacion> publicaciones = null;
-
-        try {
-            publicaciones = servicioPublicacion.listarPublicacionesPorUsuarioId(datosPublicacion.getAutorId());
-        } catch (Exception e) {
-            model.put("error", "No se encontraron publicaciones");
-        }
-
-        model.put("publicaciones", publicaciones);
-        return new ModelAndView("index-misposts", model);
-    }
     @RequestMapping(path = "/crear", method = RequestMethod.GET)
-    public ModelAndView crear(HttpSession session) {
+    public ModelAndView crear() {
 
-        return new ModelAndView("new-post");
+        ModelMap model = this.iniciarModel();
+        
+        model.put("publicacionDto", new PublicacionDto());
+        model.put("mascotas", this.servicioMascota.listarMascotasAPublicar(this.userAuth));
+
+        return new ModelAndView("new-post", model);
     }
+
+
+    @RequestMapping(path = "/publicar", method = RequestMethod.POST)
+    public ModelAndView guardarPublicacion(@ModelAttribute("publicacionDto") PublicacionDto publicacionDto, HttpServletRequest request) {
+        ModelMap model = this.iniciarModel();
+
+        Long p_id = servicioPublicacion.guardarPublicacion(publicacionDto);
+
+        if(p_id == null) {
+            model.put("error", this.servicioPublicacion.getErrorMessage());
+            return new ModelAndView("new-post",model);
+        }
+
+        return new ModelAndView("redirect: " + request.getContextPath() + "/home/mispublicaciones?pid=" + p_id);
+
+
+    }
+
+    @RequestMapping(path = "/ver", method = RequestMethod.GET)
+    public ModelAndView verPublicacion(@RequestParam Long pid, @RequestParam(required = false) String msj_response){
+        ModelMap model = iniciarModel();
+
+        model.put("publicacion", this.servicioPublicacion.findPublicacion(pid));
+
+        model.put("mensajes", this.servicioMesnaje.listarMensajesPublicacion(pid));
+
+        model.put("mensajeDto", new MensajeDto());
+
+        model.put("msj_response", msj_response);
+
+        return new ModelAndView("post-details", model);
+    }
+
+
+
+
 
 
 }
