@@ -1,17 +1,30 @@
 package ar.edu.unlam.tallerweb1.domain.publicaciones;
 
 import ar.edu.unlam.tallerweb1.model.Publicacion;
+import ar.edu.unlam.tallerweb1.model.PublicacionMensajes;
+import ar.edu.unlam.tallerweb1.model.Publicacion_favorito;
 
 import java.util.List;
 
 public interface IRepositorioPublicacion {
     Publicacion buscarPublicacionPorId(Long id);
-    Publicacion guardarPublicacion(Publicacion publicacion);
+    void guardarPublicacion(Publicacion publicacion);
+
     void modificarPublicacion(Publicacion publicacion);
 
     void eliminarPublicacion(Long IdPublicacion);
 
-    List<Publicacion> listarPublicaciones();
+    List<Publicacion> listarPublicaciones(String state);
 
     List<Publicacion> listarPublicacionesPorUsuarioId(Long idUsuario);
+
+    List<PublicacionMensajes> listarPublicacionesConMensajesPorUsuarioId(Long idUsuario);
+
+    void agregarFavorito(Publicacion_favorito favorito);
+
+    void eliminarFavorito(Publicacion_favorito favorito);
+
+    List<Publicacion_favorito> ListarFavoritosDeUsuario(Long idUsuario);
+
+
 }
