@@ -1,7 +1,9 @@
 package ar.edu.unlam.tallerweb1.domain.publicaciones;
 
+import ar.edu.unlam.tallerweb1.delivery.PublicacionDto;
+import ar.edu.unlam.tallerweb1.domain.archivos.IServicioArchivo;
 import ar.edu.unlam.tallerweb1.infrastructure.RepositorioPublicacion;
-import ar.edu.unlam.tallerweb1.model.Publicacion;
+import ar.edu.unlam.tallerweb1.model.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,11 +14,14 @@ import java.util.List;
 @Transactional
 public class ServicioPublicacion implements IServicioPublicacion{
 
+    private String errorMessage;
+    private final IServicioArchivo servicioArchivo;
     private RepositorioPublicacion repositorioPublicacion;
 
     @Autowired
-    public ServicioPublicacion(RepositorioPublicacion repositorioPublicacion){
+    public ServicioPublicacion(RepositorioPublicacion repositorioPublicacion, IServicioArchivo servicioArchivo){
         this.repositorioPublicacion = repositorioPublicacion;
+        this.servicioArchivo = servicioArchivo;
     }
 
     @Override
@@ -25,8 +30,52 @@ public class ServicioPublicacion implements IServicioPublicacion{
     }
 
     @Override
-    public Publicacion  guardarPublicacion(Publicacion publicacion) {
-       return repositorioPublicacion.guardarPublicacion(publicacion);
+    public Long  guardarPublicacion(PublicacionDto publicacionDto) {
+
+        if (!this.validarDatos(publicacionDto))
+            return null;
+        return this.publicar(publicacionDto);
+    }
+
+    private boolean validarDatos(PublicacionDto pd) {
+        if(pd.getMascota_id() == null){
+            this.errorMessage = "Debe seleccionar una mascota";
+            return false;
+        }
+        if(pd.getDireccion() == null){
+            this.errorMessage = "Debe especificar una direcci�n de entrega";
+            return false;
+        }
+
+        return true;
+    }
+
+    private Long publicar(PublicacionDto publicacionDto) {
+        Publicacion pub = this.setearPublicacion(publicacionDto);
+
+        repositorioPublicacion.guardarPublicacion(pub);
+
+        this.servicioArchivo.subirImagenesPost(publicacionDto.getFiles(), pub);
+
+        return pub.getId();
+
+    }
+
+    private Publicacion setearPublicacion(PublicacionDto publicacionDto) {
+        Publicacion p = new Publicacion();
+        p.setBio(publicacionDto.getBio());
+        p.setMascota(new Mascota(publicacionDto.getMascota_id()));
+        p.setDireccion(publicacionDto.getDireccion());
+        p.setDisponibilidad(publicacionDto.getDisponibilidad());
+        p.setLatitud(publicacionDto.getLatitud());
+        p.setLongitud(publicacionDto.getLongitud());
+        p.setDireccion(publicacionDto.getDireccion());
+        p.setProvincia(publicacionDto.getProvincia());
+        p.setCiudad(publicacionDto.getCiudad());
+
+        p.setEstado("disponible");
+
+        return p;
     }
 
     @Override
@@ -41,11 +90,52 @@ public class ServicioPublicacion implements IServicioPublicacion{
 
     @Override
     public List<Publicacion> listarPublicacionesPorUsuarioId(Long idUsuario) {
-        return repositorioPublicacion.listarPublicacionesPorUsuarioId(idUsuario);
+        return this.repositorioPublicacion.listarPublicacionesPorUsuarioId(idUsuario);
     }
 
     @Override
-    public List<Publicacion> listarPublicaciones() {
-        return repositorioPublicacion.listarPublicaciones();
+    public List<PublicacionMensajes> listarPublicacionesMensajesPorUsuarioId(Long idUsuario) {
+        return this.repositorioPublicacion.listarPublicacionesConMensajesPorUsuarioId(idUsuario);
     }
+
+    @Override
+    public List<Publicacion_favorito> listarFavoritosDeUsuario(Long idUsuario) {
+        return this.repositorioPublicacion.ListarFavoritosDeUsuario(idUsuario);
+    }
+
+    @Override
+    public void agregarFavorito(Long idPublicacion, Usuario usuario) {
+
+        Publicacion p = new Publicacion();
+
+        p.setId(idPublicacion);
+
+        Publicacion_favorito pf = new Publicacion_favorito(usuario, p);
+
+        this.repositorioPublicacion.agregarFavorito(pf);
+
+    }
+
+    @Override
+    public void eliminarFavorito(Long idPublicacion, Usuario usuario) {
+        Publicacion p = new Publicacion();
+
+        p.setId(idPublicacion);
+
+        Publicacion_favorito pf = new Publicacion_favorito(usuario, p);
+
+        this.repositorioPublicacion.eliminarFavorito(pf);
+    }
+
+    @Override
+    public List<Publicacion> listarPublicacionesDisponibles() {
+        return repositorioPublicacion.listarPublicaciones("disponible");
+    }
+
+    @Override
+    public String getErrorMessage(){
+        return this.errorMessage;
+    }
+
+
 }
