@@ -3,11 +3,11 @@ package ar.edu.unlam.tallerweb1.model;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
+import java.sql.Date;
 import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 public class Usuario {
@@ -18,12 +18,17 @@ public class Usuario {
     private String nombre;
     private String email;
     private String password;
+    private Date f_nac;
     private String domicilio;
+    private String ciudad;
+    private String provincia;
+    private Long lat;
+    private Long lng;
     private String tel;
-
-    private String rol;
-
     private String imagen;
+
+    @Column(length = 255)
+    private String presentacion;
 
     @CreationTimestamp
     private Timestamp created_at;
@@ -32,6 +37,9 @@ public class Usuario {
     private Timestamp update_at;
 
     private boolean updated = false;
+
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "usuario")
+    List<Mascota> mascotas = new ArrayList<>();
 
 
     public Usuario() {
@@ -76,12 +84,52 @@ public class Usuario {
         this.password = password;
     }
 
+    public Date getF_nac() {
+        return f_nac;
+    }
+
+    public void setF_nac(Date f_nac) {
+        this.f_nac = f_nac;
+    }
+
     public String getDomicilio() {
         return domicilio;
     }
 
     public void setDomicilio(String dni) {
         this.domicilio = dni;
+    }
+
+    public String getCiudad() {
+        return ciudad;
+    }
+
+    public void setCiudad(String ciudad) {
+        this.ciudad = ciudad;
+    }
+
+    public String getProvincia() {
+        return provincia;
+    }
+
+    public void setProvincia(String provincia) {
+        this.provincia = provincia;
+    }
+
+    public Long getLat() {
+        return lat;
+    }
+
+    public void setLat(Long lat) {
+        this.lat = lat;
+    }
+
+    public Long getLng() {
+        return lng;
+    }
+
+    public void setLng(Long lng) {
+        this.lng = lng;
     }
 
     public String getTel() {
@@ -99,12 +147,29 @@ public class Usuario {
     public void setImagen(String imagen) {
         this.imagen = imagen;
     }
+
+    public String getPresentacion() {
+        return presentacion;
+    }
+
+    public void setPresentacion(String presentacion) {
+        this.presentacion = presentacion;
+    }
+
     public Timestamp getCreated_at() {
         return created_at;
     }
 
     public void setCreated_at(Timestamp created_at) {
         this.created_at = created_at;
+    }
+
+    public Timestamp getUpdate_at() {
+        return update_at;
+    }
+
+    public void setUpdate_at(Timestamp update_at) {
+        this.update_at = update_at;
     }
 
     public boolean isUpdated() {
@@ -115,11 +180,11 @@ public class Usuario {
         this.updated = updated;
     }
 
-    public String getRol() {
-        return rol;
+    public List<Mascota> getMascotas() {
+        return mascotas;
     }
 
-    public void setRol(String rol) {
-        this.rol = rol;
+    public void setMascotas(List<Mascota> mascotas) {
+        this.mascotas = mascotas;
     }
 }
