@@ -1,67 +1,112 @@
 package ar.edu.unlam.tallerweb1.delivery;
 
-import ar.edu.unlam.tallerweb1.domain.mascota.IServicioMascota;
-import ar.edu.unlam.tallerweb1.domain.mascota.ServicioIngresarMascota;
+import ar.edu.unlam.tallerweb1.domain.auth.ServicioAuth;
+import ar.edu.unlam.tallerweb1.domain.mascota.ServicioMascota;
+import ar.edu.unlam.tallerweb1.model.Mascota;
+import ar.edu.unlam.tallerweb1.model.Usuario;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
 import org.springframework.web.servlet.ModelAndView;
-
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+import java.util.ArrayList;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+@RunWith(MockitoJUnitRunner.class)
 public class ControladorMascotaTest {
 
-
-    private IServicioMascota iServicioMascota;
+    @Mock
+    private ServicioMascota servicioMascota;
+    @Mock
+    private ServicioAuth servicioAuth;
+    @Mock
+    HttpSession session;
+    @Mock
+    HttpServletRequest request;
+    @InjectMocks
     private ControladorMascota controladorMascota;
 
-    private DatosIngresoMascota datosIngresoMascota;
-
+    private Usuario userAuth;
 
     @Before
-    public void init()
-    {
-
-        this.datosIngresoMascota  = new DatosIngresoMascota();
-        this.iServicioMascota = mock(ServicioIngresarMascota.class);
-        this.controladorMascota = new ControladorMascota(this.iServicioMascota);
+    public void init(){
+        this.userAuth = new Usuario("Ariel", "ariel@ariel", "12345");
     }
 
 
     @Test
-    public void alIngresarDatosMinimosRequeridosDeMascotaElIngresoEsExitosoYMeLlevaAMiPerfil()
-    {
-        dadoQueNoExisteMascota(this.datosIngresoMascota, true);
-        ModelAndView mav = cuandoIngresoLaMascota(this.datosIngresoMascota);
+    public void alIngresarMascotaDesdeMiPefilAlfinalizarElIngresoMellevaNuevamenteAMiPerfil() {
+        MascotaDto mascota = dadoQueExisteMascota();
+        ModelAndView mav = cuandoIngresoLaMascota(mascota);
         entoncesElIngresoEsExitoso(mav);
     }
 
     @Test
-    public void alNoIngresarDatosMinimosDeMascotaNoSePuedeIgresarLaMisma()
+    public void alNoIngresarLosDatosRequeridosDeLaMascotaNoSePuedeGuardarLaMisma()
     {
-
-        dadoQueNoExisteMascota(this.datosIngresoMascota, false);
-        ModelAndView mav = cuandoIngresoLaMascota(this.datosIngresoMascota);
+        MascotaDto mascota = dadoQueTengoUnaMascotaIncompleta();
+        ModelAndView mav = cuandoIngresoLaMascota(mascota);
         entoncesElIngresoNoEsExitoso(mav);
-    }
 
-    private void entoncesElIngresoEsExitoso(ModelAndView mav) {
-        assertThat(mav.getViewName()).isEqualTo("profile");
-        assertThat(mav.getModel().get("msg")).isEqualTo("Mascota Ingresada");
     }
 
     private void entoncesElIngresoNoEsExitoso(ModelAndView mav) {
-        assertThat(mav.getViewName()).isEqualTo("new-mascot");
-        assertThat(mav.getModel().get("msg")).isEqualTo("No se Pudo Ingresar La mascota, ingrese los campos mínimos");
+        assertThat(mav.getViewName()).isEqualTo( "new-mascot");
     }
 
-    private ModelAndView cuandoIngresoLaMascota(DatosIngresoMascota datosIngresoMascota) {
-        return controladorMascota.ingresarMascota(datosIngresoMascota);
+    private MascotaDto dadoQueTengoUnaMascotaIncompleta() {
+
+        MascotaDto mascotadto = new MascotaDto();
+
+
+
+        String target = "publicacion";
+        when(session.getAttribute("target")).thenReturn(target);
+        when(servicioMascota.guardar(mascotadto,this.userAuth)).thenReturn(null);
+
+
+
+        return mascotadto;
+    }
+
+    private MascotaDto dadoQueExisteMascota() {
+
+
+        MascotaDto mascotadto = new MascotaDto();
+
+        String target = "perfil";
+        when(session.getAttribute("target")).thenReturn(target);
+
+
+
+        return mascotadto;
+
+    }
+
+    private ModelAndView cuandoIngresoLaMascota(MascotaDto mascotadto) {
+
+        when(servicioAuth.getUsuarioAutenticado()).thenReturn(userAuth);
+        return controladorMascota.guardarMascota(mascotadto, session, request);
+
+    }
+
+    private void entoncesElIngresoEsExitoso(ModelAndView mav) {
+        assertThat(mav.getViewName()).isEqualTo( "redirect: " + request.getContextPath() + "/perfil/actividad/mascotas");
+
     }
 
 
-    private void dadoQueNoExisteMascota(DatosIngresoMascota datosIngresoMascota, Boolean retorno) {
-        when(this.iServicioMascota.sonValidos(datosIngresoMascota.getNombre(),datosIngresoMascota.getTipo(), datosIngresoMascota.getRaza(), datosIngresoMascota.getPeso(),datosIngresoMascota.getNacimiento(),datosIngresoMascota.getObs(),datosIngresoMascota.getFoto())).thenReturn(retorno);
+    private void dadoQueExisteMascotas(Usuario usuario) {
+
+        when(servicioMascota.listarMascotaPorUsuario(usuario)).thenReturn(new ArrayList<Mascota>());
 
     }
+
+
+
+
 }
