@@ -1,6 +1,6 @@
 package ar.edu.unlam.tallerweb1.delivery;
 
-import ar.edu.unlam.tallerweb1.model.Publicacion;
+
 import ar.edu.unlam.tallerweb1.domain.publicaciones.ServicioPublicacion;
 import ar.edu.unlam.tallerweb1.model.Usuario;
 import org.junit.Test;
@@ -10,6 +10,7 @@ import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 
@@ -24,29 +25,27 @@ public class ControladorPublicacionTest {
     @Mock
     HttpSession session;
 
+    @Mock
+    HttpServletRequest request;
+
     @InjectMocks
     ControladorPublicacion controladorPublicacion;
 
-    @Test
-    public void alPedirTodasLasPublicacionesDevuelveLaListaCompleta() {
-        dadoQueExistenPublicaciones();
-        ModelAndView vista = cuandoQuieroVerTodasLasPublicaciones();
-        entoncesMeDevuelveLaVistaCorrecta(vista);
-    }
+
 
     @Test
     public void testAlGuardarPublicacion() {
-        DatosPublicacion publicacion = dadoQueTengoDatosDePublicacionDeUnUsuario();
+        PublicacionDto publicacion = dadoQueTengoDatosDePublicacionDeUnUsuario();
         ModelAndView vista = cuandoQuieroGuardarPublicacion(publicacion);
         entoncesMeDevuelveLaVistaCorrecta(vista);
     }
 
 
 
-    private DatosPublicacion dadoQueTengoDatosDePublicacionDeUnUsuario() {
-        DatosPublicacion datosPublicacion = new DatosPublicacion();
-        datosPublicacion.setTitulo("Prueba");
-        datosPublicacion.setCuerpo("Esto es un Test de publicacion");
+    private PublicacionDto dadoQueTengoDatosDePublicacionDeUnUsuario() {
+        PublicacionDto publicacionDto = new PublicacionDto();
+
+        publicacionDto.setBio("Esto es un Test de publicacion");
 
         Usuario usuario = new Usuario("juani","juani@aa","1234");
 
@@ -55,27 +54,19 @@ public class ControladorPublicacionTest {
         when(session.getAttribute("usuarioAutenticado")).thenReturn(usuario);
 
 
-        return datosPublicacion;
+        return publicacionDto;
     }
-    private ModelAndView cuandoQuieroGuardarPublicacion(DatosPublicacion datosPublicacion) {
+    private ModelAndView cuandoQuieroGuardarPublicacion(PublicacionDto publicacionDto) {
 
-
-        when(servicioPublicacion.guardarPublicacion(new Publicacion())).thenReturn(null);
-
-
-        return controladorPublicacion.guardarPublicacion(datosPublicacion, session);
+        return controladorPublicacion.guardarPublicacion(publicacionDto, request );
     }
 
     private void dadoQueExistenPublicaciones() {
 
-        when(servicioPublicacion.listarPublicaciones()).thenReturn(new ArrayList<Publicacion>());
+        when(servicioPublicacion.listarPublicacionesDisponibles()).thenReturn(new ArrayList<>());
 
     }
 
-    private ModelAndView cuandoQuieroVerTodasLasPublicaciones() {
-
-        return controladorPublicacion.verPublicacionesFeed();
-    }
 
     private static void entoncesMeDevuelveLaVistaCorrecta(ModelAndView vista) {
         assertThat(vista.getViewName()).isEqualTo("index-feed");
